Refractor all c# code

+++ ICS.Project.App/ViewModels/MessengerScreenViewModels/TeamDetailViewModel.cs
@@ -13,25 +13,8 @@ namespace ICS.Project.App.ViewModels.MessengerScreenViewModels
 {
     public class TeamDetailViewModel : ViewModelBase, IViewModel
     {
-        private readonly IUsersRepository _usersRepository;
         private readonly ITeamsRepository _teamsRepository;
-
-        public bool IsDescriptionInEditMode { get; set; }
-        public UserModel SelectedUserInComboBox { get; set; }
-
-        public ObservableCollection<UserModel> Members { get; set; } = new ObservableCollection<UserModel>();
-        public ObservableCollection<UserModel> AvailableMembers { get; set; } = new ObservableCollection<UserModel>();
-        public TeamModel Team { get; set; }
-        public UserModel LoggedUser { get; set; }
-
-
-        public ICommand EditDescriptionCommand { get; set; }
-        public ICommand SaveDescriptionCommand { get; set; }
-        public ICommand MemberClickedCommand { get; set; }
-        public ICommand RemoveMemberCommand { get; set; }
-        public ICommand AddNewMemberCommand { get; set; }
-        public ICommand DeleteTeamCommand { get; set; }
-
+        private readonly IUsersRepository _usersRepository;
 
         public TeamDetailViewModel(IUsersRepository usersRepository, ITeamsRepository teamsRepository)
         {
@@ -50,10 +33,29 @@ namespace ICS.Project.App.ViewModels.MessengerScreenViewModels
             Mediator.Instance.Register<UserLogoutMessage>(UserLogout);
         }
 
+        public ICommand EditDescriptionCommand { get; set; }
+        public ICommand SaveDescriptionCommand { get; set; }
+        public ICommand MemberClickedCommand { get; set; }
+        public ICommand RemoveMemberCommand { get; set; }
+        public ICommand AddNewMemberCommand { get; set; }
+        public ICommand DeleteTeamCommand { get; set; }
+
+        public bool IsDescriptionInEditMode { get; set; }
+        public UserModel SelectedUserInComboBox { get; set; }
+        public ObservableCollection<UserModel> Members { get; set; } = new ObservableCollection<UserModel>();
+        public ObservableCollection<UserModel> AvailableMembers { get; set; } = new ObservableCollection<UserModel>();
+        public TeamModel Team { get; set; }
+        public UserModel LoggedUser { get; set; }
+
+        public void Load()
+        {
+            IsDescriptionInEditMode = false;
+        }
+
 
         private void MemberClicked(UserModel clickedMember)
         {
-            Mediator.Instance.Send(new UserWasClickedMessage { User = clickedMember });
+            Mediator.Instance.Send(new UserWasClickedMessage {User = clickedMember});
         }
 
         private void UserLogged(UserLoggedMessage userLoggedMessage)
@@ -82,24 +84,14 @@ namespace ICS.Project.App.ViewModels.MessengerScreenViewModels
                 Team = selectedTeamMessage.Team;
 
                 Members.Clear();
-                foreach (var member in _teamsRepository.GetTeamMembers(Team.ID))
-                {
-                    Members.Add(member);
-                }
+                foreach (var member in _teamsRepository.GetTeamMembers(Team.ID)) Members.Add(member);
 
                 AvailableMembers.Clear();
                 foreach (var member in _usersRepository.GetAll().Where(s => !Members.Select(e => e.ID).Contains(s.ID)))
-                {
                     AvailableMembers.Add(member);
-                }
 
                 SelectedUserInComboBox = AvailableMembers.FirstOrDefault();
             }
-        }
-
-        public void Load()
-        {
-            IsDescriptionInEditMode = false;
         }
 
         public void AddNewTeamMember()
@@ -121,17 +113,20 @@ namespace ICS.Project.App.ViewModels.MessengerScreenViewModels
             if (clickedUser.ID == LoggedUser.ID)
             {
                 if (Members.Count() == 1)
-                {
-                    messageBoxResult = MessageBox.Show("You are the last member of this team. The team with all posts will be discarded if you leave! Are you sure to continue?", "Warning", MessageBoxButton.YesNo, MessageBoxImage.Warning, MessageBoxResult.No);
-                }
+                    messageBoxResult = MessageBox.Show(
+                        "You are the last member of this team. The team with all posts will be discarded if you leave! Are you sure to continue?",
+                        "Warning", MessageBoxButton.YesNo, MessageBoxImage.Warning, MessageBoxResult.No);
                 else
-                {
-                    messageBoxResult = MessageBox.Show("Are you sure to remove yourself from this team? All your posts will reamain, but you lose access to the team!", "Warning", MessageBoxButton.YesNo, MessageBoxImage.Warning, MessageBoxResult.No);
-                }
+                    messageBoxResult = MessageBox.Show(
+                        "Are you sure to remove yourself from this team? All your posts will reamain, but you lose access to the team!",
+                        "Warning", MessageBoxButton.YesNo, MessageBoxImage.Warning, MessageBoxResult.No);
             }
             else
             {
-                messageBoxResult = MessageBox.Show($"Are you sure to cancel {clickedUser.FullName} access to this team? All posts from the user will reamain.", "Warning", MessageBoxButton.YesNo, MessageBoxImage.Warning, MessageBoxResult.No);
+                messageBoxResult =
+                    MessageBox.Show(
+                        $"Are you sure to cancel {clickedUser.FullName} access to this team? All posts from the user will reamain.",
+                        "Warning", MessageBoxButton.YesNo, MessageBoxImage.Warning, MessageBoxResult.No);
             }
 
             if (messageBoxResult == MessageBoxResult.No) return;
@@ -140,11 +135,8 @@ namespace ICS.Project.App.ViewModels.MessengerScreenViewModels
 
             if (clickedUser.ID == LoggedUser.ID)
             {
-                if (Members.Count() == 1)
-                {
-                    _teamsRepository.RemoveWithAllPostsAndComments(Team.ID);
-                }
-                Mediator.Instance.Send(new UserLostAccessToTeam { Team = Team });
+                if (Members.Count() == 1) _teamsRepository.RemoveWithAllPostsAndComments(Team.ID);
+                Mediator.Instance.Send(new UserLostAccessToTeam {Team = Team});
             }
             else
             {
@@ -156,12 +148,14 @@ namespace ICS.Project.App.ViewModels.MessengerScreenViewModels
 
         private void DeleteTeam()
         {
-            MessageBoxResult messageBoxResult = MessageBox.Show($"Team {Team.Name} with all posts will be deleted! Are you sure to continue?", "Warning", MessageBoxButton.YesNo, MessageBoxImage.Warning, MessageBoxResult.No); 
+            var messageBoxResult =
+                MessageBox.Show($"Team {Team.Name} with all posts will be deleted! Are you sure to continue?",
+                    "Warning", MessageBoxButton.YesNo, MessageBoxImage.Warning, MessageBoxResult.No);
 
             if (messageBoxResult == MessageBoxResult.No) return;
 
             _teamsRepository.RemoveWithAllPostsAndComments(Team.ID);
-            Mediator.Instance.Send(new UserLostAccessToTeam { Team = Team });
+            Mediator.Instance.Send(new UserLostAccessToTeam {Team = Team});
         }
